complete

+++ action/delete_chore_action.php
@@ -3,35 +3,42 @@ include '../settings/connection.php';
 
 $response = array('success' => false, 'message' => '');
 
-if (isset($_GET['id'])) {
-    $choreId = $_GET['id'];
+include_once('../settings/core.php');
 
-    $sql = "DELETE FROM Chores WHERE cid = ?";
+if (checkUserRole() != 1) {
+    $response['message'] = "Error: You do not have permission to use this feature.";
+    echo json_encode($response);
+} else {
 
-    $stmt = $conn->prepare($sql);
+    if (isset($_GET['id'])) {
+        $choreId = $_GET['id'];
 
-    if ($stmt) {
-        $stmt->bind_param("i", $choreId);
-        $stmt->execute();
+        $sql = "DELETE FROM Chores WHERE cid = ?";
 
-        if ($stmt->affected_rows > 0) {
-            $response['success'] = true;
-            $response['message'] = "Chore deleted successfully.";
+        $stmt = $conn->prepare($sql);
+
+        if ($stmt) {
+            $stmt->bind_param("i", $choreId);
+            $stmt->execute();
+
+            if ($stmt->affected_rows > 0) {
+                $response['success'] = true;
+                $response['message'] = "Chore deleted successfully.";
+            } else {
+                $response['success'] = false;
+                $response['message'] = "Error: Unable to delete chore. Please try again.";
+            }
+
+            $stmt->close();
         } else {
             $response['success'] = false;
-            $response['message'] = "Error: Unable to delete chore. Please try again.";
+            $response['message'] = "Error: Unable to prepare statement. Please try again.";
         }
-
-        $stmt->close();
     } else {
         $response['success'] = false;
-        $response['message'] = "Error: Unable to prepare statement. Please try again.";
+        $response['message'] = "Error: Wrong request method. Please try again.";
     }
-} else {
-    $response['success'] = false;
-    $response['message'] = "Error: Wrong request method. Please try again.";
-}
 
-$conn->close();
-echo json_encode($response);
-?>
+    $conn->close();
+    echo json_encode($response);
+}
